+++ src/main/java/com/koc/user/domain/User.java
@@ -1,6 +1,5 @@
 package com.koc.user.domain;
 
-import com.koc.user.application.dto.UserDto;
 import com.koc.user.adapter.out.persistence.UserEntity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
