+++ src/main/java/com/koc/user/domain/UserTokenDto.java
@@ -1,4 +1,4 @@
-package com.koc.user.application.dto;
+package com.koc.user.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
